refactor: move src/misc/global-state.ts -> src/core/state.ts
