+++ src/core/state.ts
