+++ src/rendering/renderer-codeblock.ts
@@ -12,7 +12,7 @@ import Callbag, { flatMap, fromEvent, map, pairwise, takeUntil } from 'src/utili
 import { CodeBlockSettingsDialog } from 'src/settings/dialogs'
 import { isObjectEmpty } from 'src/utilities/utilities'
 import { TabRow } from 'src/workspace/db-schema'
-import { globalState } from 'src/misc/global-state'
+import { globalState } from 'src/core/state'
 
 
 export type CodeBlockRenderer = ReturnType<typeof CodeBlockRenderer>
